updaters and submobjects initialized at None
It turns out that selecting as default an empty list as an attribute
with attrs seems like a bad idea. Indeed the said list(reference)
will be the same for every instance of the class.
This issue may arise again with other default attributes with objects
whose state will change.