Added sanity checks to :meth:`~.Mobject.add_to_back` for Mobjects (#1220)
* Add Mobject add_to_back sanity checks

- Raises ValueError when Mobject tries to add itself
- Raises TypeError when a non-Mobject is added
- Filters out incoming duplicate submobjects if at least one instance of
that submobject exists in the list

* Filter incoming Mobjects into a set before adding

First, incoming list of Mobjects will be filtered into a set with no
duplicates. Then, any mobject in submobjects that is a duplicate of
an incoming mobject will be removed. The filtered set of mobjects will
be added to back and the remaining submobjects are appended.

* Add isort to pre-commit config

* Isort test_vectorized_mobject.py

* Was already added

* Update .pre-commit-config.yaml

Co-authored-by: kolibril13 <44469195+kolibril13@users.noreply.github.com>
Co-authored-by: Benjamin Hackl <devel@benjamin-hackl.at>