Always scale text according to the screen size
This should help people because they don't need to scale
every time instead it will automatically scale.

It's an initial implementation, only for Text now. It
can be extended to other implementations also.

This also fixes #1327
by adding an option to disable word wraps.

Also update link to ManimPango wheels
and update tests data as there is a small
change in SVGMobject which needs to regenerate the data.

Signed-off-by: Naveen M K <naveen@syrusdark.website>