Added tests for resolution flag
Added test cases to make sure the resolution flag does what its supposed
to. It queues up a handful of common 16:9 and a couple 4:3 resolutions
and tries to render all of them with all acceptable separators. Once
they render it checks to make sure that the rendered resolution matches
the requested resolution.