Changed log_to_file mechanism
Previously log_to_file was called within the _config module.
That caused problems because log_to_file requires the name of the scene
as well which gets initialised to None.
Now we set the logger within scene_file_writer

The set_file_logger had the ability to change the change config.verbosity,
I have removed that ability as well, as it felt unnecessary.