Add assertion if font is not found for Text, Code, and MarkupText
This adds tests for this behavior and makes sure that if a font is registered via register_font, the assertions work correctly.

If a font is not specified, we do not assert, as this will cause Manim to use a default font.

Some tests needed to be updated to use cross-platform font names. Monospace and Sans are found on both my Windows and Linux environment, so that seems like a reasonable thing to try.