feat(lib): add basic VideoMobject class
Hello, I actually implemented this for a personal need, but I guess other people could enjoy such feature, as in #760.

The current implementation is very minimalist, and may be improved, or tested / better documented.

If OpenCV is added as a dependency, then we could also directly load videos from video files, which could be a nice feature.