Fix incorrect assignment in Mobject.put_start_and_end_on
PR #3718 changed the behavior of `Mobject.put_start_and_end_on` when
`start == end`, however the assigment to `self.points` was
incorrect. The existing code assigned the Point3D directly to
`self.points`, that then becomes an array with shape `(3,)`, while
instead it should really have shape `(1, 3)`.

PR #4027 added a comment that this should be fixed, but did not fix
the issue. The comment is now superfluous.

This commit fixes the incorrect code and associated test.